requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
langchain-community>=0.2.0
tavily-python>=0.3.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
streamlit>=1.35.0
plotly>=5.18.0
langgraph-cli[inmem]>=0.1.0
